1、添加乐嗨嗨渠道资源 2、添加弹框提示 3、打包完成后，添加自动跳转目标apk目录操作

+++ RongSDKInjectTool/src/com/rong/utils/FileUtils.java
@@ -103,16 +103,18 @@ public class FileUtils {
 		if (!dirFile.exists()) {
 			return;
 		}
-		File[] files = dirFile.listFiles();
-		for (int i = 0; i < files.length; i++) {
-			File file = files[i];
-			if (file.isDirectory()) {
-				rmdir(file);
-			} else {
-				try {
-					file.delete();
-				} catch (Exception e) {
-					e.printStackTrace();
+		if (dirFile.isDirectory()) {
+			File[] files = dirFile.listFiles();
+			for (int i = 0; i < files.length; i++) {
+				File file = files[i];
+				if (file.isDirectory()) {
+					rmdir(file);
+				} else {
+					try {
+						file.delete();
+					} catch (Exception e) {
+						e.printStackTrace();
+					}
 				}
 			}
 		}
